backend/pyproject.toml
[tool.poetry]
name = "foodflex"
version = "0.1.0"
description = ""
authors = ["AishaFathalla <156489851+AishaFathalla@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
Django = "^3.2"
djangorestframework = "^3.12.4"
djongo = "^1.3.6"
pymongo = "^4.10.1"
django-environ = "^0.11.2"
django-cors-headers = "^4.4.0"
mongoengine = "^0.29.1"
djangorestframework-simplejwt = "^5.3.1"

[tool.poetry.group.dev.dependencies]
djongo = "^1.3.6"
pymongo = "^4.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
